% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcd.R
\name{calculate_gcd}
\alias{calculate_gcd}
\title{This is a simple function using Rcpp that performs Euclides algorithm}
\usage{
calculate_gcd(a, b)
}
\arguments{
\item{a}{numeric}

\item{b}{numeric}
}
\description{
This is a simple function using Rcpp that performs Euclides algorithm
}
