% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gcd}
\alias{gcd}
\title{This is a simple function using Rcpp that performs Euclides algorithm}
\usage{
gcd(a, b)
}
\arguments{
\item{a}{numeric}

\item{b}{numeric}
}
\description{
This is a simple function using Rcpp that performs Euclides algorithm
}
